refactor(ui): move authSlice to shared/auth
- `auth` 세그먼트를 작성했으니, 관련 로직이 담긴 파일을 한 곳에 모아 구조를 파악하기 쉽게 합니다.
