+++ packages/ui/src/shared/auth/index.ts
@@ -1,1 +1,2 @@
+export { authReducer, selectUserId } from './authSlice';
 export { login, logout } from './thunks';
