+++ packages/ui/src/shared/model/index.ts
@@ -1,9 +1,1 @@
-export {
-  authReducer,
-  authSlice,
-  clearAccessToken,
-  selectIsAuthenticated,
-  selectUserId,
-  setAccessToken,
-} from './authSlice';
 export { useAppDispatch, useAppSelector } from './hook';
